+++ src/edu/mit/compilers/codegen/names/VariableName.java
@@ -3,7 +3,7 @@ package edu.mit.compilers.codegen.names;
 public class VariableName extends AssignableName {
     String label;
 
-    public VariableName(String label, int size) {
+    public VariableName(String label, long size) {
         super(size);
         this.label = label;
     }
